\name{cloneRepo}
\alias{cloneRepo}
\title{Clone Repo}
\usage{
cloneRepo(repo, owner, basePath = "student_repos")
}
\description{
Clones the GitHub repository "repo", owned by "owner" to the destination "basePath".
}
\examples{
cloneRepo(repo = "Lastname_Firstname", owner = "Bucknell-Biol364", basePath = "../../student_repos")
}
